+++ src/test/kotlin/krangl/test/CoreVerbsTest.kt
@@ -6,7 +6,6 @@ import krangl.*
 import org.apache.commons.csv.CSVFormat
 
 
-val sleepData = DataFrame.fromCSV(DataFrame::class.java.getResourceAsStream("data/msleep.csv"))
 val irisData = DataFrame.fromCSV(DataFrame::class.java.getResourceAsStream("data/iris.txt"), format = CSVFormat.TDF)
 val flights = DataFrame.fromCSV(DataFrame::class.java.getResourceAsStream("data/nycflights.tsv.gz"), format = CSVFormat.TDF, isCompressed = true)